Pull detail tests out of tests.py into their own file
